+++ sql/vector.sql
@@ -105,6 +105,9 @@ CREATE FUNCTION vector_avg(double precision[]) RETURNS vector
 CREATE FUNCTION vector_combine(double precision[], double precision[]) RETURNS double precision[]
 	AS 'MODULE_PATHNAME' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
+CREATE FUNCTION vector_concat(vector, vector) RETURNS vector
+	AS 'MODULE_PATHNAME' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
+
 -- vector aggregates
 
 CREATE AGGREGATE avg(vector) (
@@ -192,6 +195,10 @@ CREATE OPERATOR - (
 CREATE OPERATOR * (
 	LEFTARG = vector, RIGHTARG = vector, PROCEDURE = vector_mul,
 	COMMUTATOR = *
+);
+
+CREATE OPERATOR || (
+	LEFTARG = vector, RIGHTARG = vector, PROCEDURE = vector_concat
 );
 
 CREATE OPERATOR < (